add validation that user must be authenticated to delete their account

+++ part4/blog-list-app-backend/controllers/users.js
@@ -2,6 +2,7 @@ const bcrypt = require('bcrypt');
 const usersRouter = require('express').Router();
 
 const User = require('../models/user');
+const { decodeToken } = require('../utils/authentication');
 
 usersRouter.get('/', async (_req, res) => {
   const users = await User.find({}).populate('blogs', { title: 1, url: 1, author: 1 });
@@ -43,6 +44,19 @@ usersRouter.post('/', async (req, res) => {
 
 usersRouter.delete('/:id', async (req, res) => {
   const id = req.params.id;
+
+  const decodedToken = decodeToken(req.token);
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  if (id !== user._id.toString()) {
+    return res.status(401).json({ error: 'not authorized to perform this operation' });
+  }
+
   await User.findByIdAndRemove(id);
 
   res.status(204).end();
